Adding Hamiltonian class
